Return type of the Usb interrupt write changed to std::size_t (#123).

+++ include/com/UsbComm.h
@@ -40,7 +40,7 @@ namespace plug::com
         bool isOpen() const;
 
         template <class Container>
-        std::int32_t interruptWrite(std::uint8_t endpoint, Container c)
+        std::size_t interruptWrite(std::uint8_t endpoint, Container c)
         {
             return interruptWriteImpl(endpoint, c.data(), c.size());
         }
@@ -48,7 +48,7 @@ namespace plug::com
         std::vector<std::uint8_t> interruptReceive(std::uint8_t endpoint, std::size_t recvSize);
 
     private:
-        std::int32_t interruptWriteImpl(std::uint8_t endpoint, std::uint8_t* data, std::size_t size);
+        std::size_t interruptWriteImpl(std::uint8_t endpoint, std::uint8_t* data, std::size_t size);
 
         void initInterface();
